# swagger.yaml
openapi: 3.0.0
info:
  title: Jobs API
  contact: {}
  version: '1.0'
servers:
  - url: https://mentordeployment.onrender.com/api/v1
    variables: {}
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register User
      operationId: RegisterUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
            example:
              name: bob
              email: bob@gmail.com
              password: secret
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login User
      operationId: LoginUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserRequest'
            example:
              email: susan@gmail.com
              password: secret
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /jobs:
    post:
      tags:
        - Jobs
      summary: Create Job
      operationId: CreateJob
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
            example:
              company: google
              position: front-end developer
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
        - Jobs
      summary: Get All Jobs
      operationId: GetAllJobs
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /jobs/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The job ID
    get:
      tags:
        - Jobs
      summary: Get Single Job
      operationId: GetSingleJob
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false

    patch:
      tags:
        - Jobs
      summary: Update Job
      operationId: UpdateJob
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJobRequest'
            example:
              company: random
              position: intern
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
        - Jobs
      summary: Delete Job
      operationId: DeleteJob
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /images/users/{userId}:
    parameters:
      - in: path
        name: userId
        schema:
          type: string
        required: true
        description: The user ID
    post:
      tags:
        - Images
      summary: Upload or replace user image
      operationId: UploadUserImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Image metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
    get:
      tags:
        - Images
      summary: Get user image binary stream
      operationId: GetUserImage
      responses:
        '200':
          description: User image bytes
          content:
            image/*:
              schema:
                type: string
                format: binary
    delete:
      tags:
        - Images
      summary: Delete user image
      operationId: DeleteUserImage
      responses:
        '200':
          description: Image deleted
  /images/jobs/{jobId}:
    parameters:
      - in: path
        name: jobId
        schema:
          type: string
        required: true
        description: The job ID
    post:
      tags:
        - Images
      summary: Upload or replace job image
      operationId: UploadJobImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Image metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
    get:
      tags:
        - Images
      summary: Get job image binary stream
      operationId: GetJobImage
      responses:
        '200':
          description: Job image bytes
          content:
            image/*:
              schema:
                type: string
                format: binary
    delete:
      tags:
        - Images
      summary: Delete job image
      operationId: DeleteJobImage
      responses:
        '200':
          description: Image deleted
  /ws:
    get:
      tags:
        - WebSocket
      summary: Get WebSocket Connection Information
      operationId: GetWebSocketInfo
      parameters: []
      responses:
        '200':
          description: WebSocket connection details and supported operations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebSocketInfoResponse'
      deprecated: false
      security: []
components:
  schemas:
    RegisterUserRequest:
      title: RegisterUserRequest
      required:
        - name
        - email
        - password
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        name: bob
        email: bob@gmail.com
        password: secret
    LoginUserRequest:
      title: LoginUserRequest
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: susan@gmail.com
        password: secret
    CreateJobRequest:
      title: CreateJobRequest
      required:
        - company
        - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: google
        position: front-end developer
    UpdateJobRequest:
      title: UpdateJobRequest
      required:
        - company
        - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: random
        position: intern
    WebSocketInfoResponse:
      title: WebSocketInfoResponse
      type: object
      properties:
        message:
          type: string
        websocket:
          type: object
          properties:
            url:
              type: string
            protocol:
              type: string
            supportedCollections:
              type: array
              items:
                type: string
            messageTypes:
              type: object
    ImageResponse:
      title: ImageResponse
      type: object
      properties:
        image:
          type: object
          properties:
            id:
              type: string
            ownerType:
              type: string
              enum:
                - user
                - job
            ownerId:
              type: string
            contentType:
              type: string
            size:
              type: number
            filename:
              type: string
            createdAt:
              type: string
            updatedAt:
              type: string
            url:
              type: string
            exampleMessages:
              type: object
        documentation:
          type: object
          properties:
            description:
              type: string
            features:
              type: array
              items:
                type: string
            collections:
              type: object
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
tags:
  - name: Auth
    description: ''
  - name: Jobs
    description: ''
  - name: WebSocket
    description: Real-time database change notifications
  - name: Images
    description: Upload and stream user/job images stored in GridFS